+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### vNEXT
 
 * add commentDescription to printSchema call to match other uses [PR #745](https://github.com/apollographql/graphql-tools/pull/745)
+* Add `createResolver` option to `makeRemoteExecutableSchema` [PR #734](https://github.com/apollographql/graphql-tools/pull/734)
 
 ### v2.24.0